refactor(sanity): move bundle metadata store to resource cache
