--- packages/sanity/src/_singletons/core/bundles/BundlesContext.ts
@@ -1,8 +1,0 @@
-import {createContext} from 'react'
-
-import type {BundlesContextValue} from '../../../core/store/bundles/BundlesProvider'
-
-/**
- * @internal
- */
-export const BundlesContext = createContext<BundlesContextValue | null>(null)
