+++ packages/sanity/src/_singletons/index.ts
@@ -1,4 +1,3 @@
-export * from './core/bundles/BundlesContext'
 export * from './core/changeIndicators/ConnectorContext'
 export * from './core/components/previewCard/PreviewCardContext'
 export * from './core/components/scroll/scrollContext'
