+++ packages/sanity/src/core/bundles/hooks/usePerspective.tsx
@@ -1,6 +1,6 @@
 import {useRouter} from 'sanity/router'
 
-import {useBundles} from '../../store/bundles/BundlesProvider'
+import {useBundles} from '../../store/bundles'
 import {type BundleDocument} from '../../store/bundles/types'
 import {LATEST} from '../util/const'
