+++ packages/sanity/src/core/releases/tool/__tests__/BundlesOverview.test.tsx
@@ -4,8 +4,7 @@ import {type ReactNode} from 'react'
 
 import {queryByDataUi} from '../../../../../test/setup/customQueries'
 import {createTestProvider} from '../../../../../test/testUtils/TestProvider'
-import {useBundlesStore} from '../../../store/bundles'
-import {BundlesProvider} from '../../../store/bundles/BundlesProvider'
+import {useBundles} from '../../../store/bundles'
 import {type BundleDocument} from '../../../store/bundles/types'
 import {releasesUsEnglishLocaleBundle} from '../../i18n'
 import BundlesOverview from '../BundlesOverview'
@@ -15,8 +14,8 @@ jest.mock('../../../store/bundles/useBundleOperations', () => ({
   useBundleOperations: jest.fn().mockReturnValue({deleteBundle: jest.fn()}),
 }))
 
-jest.mock('../../../store/bundles/useBundlesStore', () => ({
-  useBundlesStore: jest.fn(),
+jest.mock('../../../store/bundles', () => ({
+  useBundles: jest.fn(),
 }))
 
 jest.mock('sanity/router', () => ({
@@ -29,15 +28,11 @@ const createWrapper = async () => {
     resources: [releasesUsEnglishLocaleBundle],
   })
   return function Wrapper({children}: {children: ReactNode}) {
-    return (
-      <TestProvider>
-        <BundlesProvider>{children}</BundlesProvider>
-      </TestProvider>
-    )
+    return <TestProvider>{children}</TestProvider>
   }
 }
 
-const mockUseBundleStore = useBundlesStore as jest.Mock<typeof useBundlesStore>
+const mockUseBundleStore = useBundles as jest.Mock<typeof useBundles>
 
 describe('BundlesOverview', () => {
   describe('when loading bundles', () => {
@@ -45,7 +40,6 @@ describe('BundlesOverview', () => {
       mockUseBundleStore.mockReturnValue({
         data: null,
         loading: true,
-        error: null,
         dispatch: jest.fn(),
       })
 
@@ -78,7 +72,6 @@ describe('BundlesOverview', () => {
       mockUseBundleStore.mockReturnValue({
         data: [],
         loading: false,
-        error: null,
         dispatch: jest.fn(),
       })
       const wrapper = await createWrapper()
@@ -116,7 +109,6 @@ describe('BundlesOverview', () => {
       mockUseBundleStore.mockReturnValue({
         data: bundles,
         loading: false,
-        error: null,
         dispatch: jest.fn(),
       })
       const wrapper = await createWrapper()
