+++ packages/sanity/src/core/store/bundles/index.ts
@@ -1,3 +1,2 @@
-export * from './BundlesProvider'
 export * from './types'
-export {useBundlesStore} from './useBundlesStore'
+export * from './useBundles'
