+++ packages/sanity/src/core/store/bundles/reducer.ts
@@ -16,7 +16,7 @@ interface BundleUpdatedAction {
 }
 
 interface BundlesSetAction {
-  payload: BundleDocument[]
+  payload: BundleDocument[] | null
   type: 'BUNDLES_SET'
 }
 
@@ -25,23 +25,33 @@ interface BundleReceivedAction {
   type: 'BUNDLE_RECEIVED'
 }
 
+interface LoadingStateChangedAction {
+  payload: {
+    loading: boolean
+    error: Error | undefined
+  }
+  type: 'LOADING_STATE_CHANGED'
+}
+
 export type bundlesReducerAction =
   | BundleAddedAction
   | BundleDeletedAction
   | BundleUpdatedAction
   | BundlesSetAction
   | BundleReceivedAction
+  | LoadingStateChangedAction
 
 export interface bundlesReducerState {
   bundles: Map<string, BundleDocument>
+  state: 'initialising' | 'loading' | 'loaded' | 'error'
+  error?: Error
 }
 
-function createBundlesSet(bundles: BundleDocument[]) {
-  const bundlesById = bundles.reduce((acc, bundle) => {
+function createBundlesSet(bundles: BundleDocument[] | null) {
+  return (bundles ?? []).reduce((acc, bundle) => {
     acc.set(bundle._id, bundle)
     return acc
   }, new Map<string, BundleDocument>())
-  return bundlesById
 }
 
 export function bundlesReducer(
@@ -49,6 +59,14 @@ export function bundlesReducer(
   action: bundlesReducerAction,
 ): bundlesReducerState {
   switch (action.type) {
+    case 'LOADING_STATE_CHANGED': {
+      return {
+        ...state,
+        state: action.payload.loading ? 'loading' : 'loaded',
+        error: action.payload.error,
+      }
+    }
+
     case 'BUNDLES_SET': {
       // Create an object with the BUNDLE id as key
       const bundlesById = createBundlesSet(action.payload)
