+++ packages/sanity/src/core/store/bundles/types.ts
@@ -1,6 +1,10 @@
 import {type ColorHueKey} from '@sanity/color'
 import {type IconSymbol} from '@sanity/icons'
 import {type SanityDocument} from '@sanity/types'
+import {type Dispatch} from 'react'
+import {type Observable} from 'rxjs'
+
+import {type bundlesReducerAction, type bundlesReducerState} from './reducer'
 
 /**
  * @internal
@@ -22,4 +26,12 @@ export interface BundleDocument extends SanityDocument {
  */
 export function isBundleDocument(doc: unknown): doc is BundleDocument {
   return typeof doc === 'object' && doc !== null && '_type' in doc && doc._type === 'bundle'
+}
+
+/**
+ * @internal
+ */
+export interface BundlesStore {
+  state$: Observable<bundlesReducerState>
+  dispatch: Dispatch<bundlesReducerAction>
 }
